Extract composer order preview
